Split up CommandHandler class
