+++ src/main/java/net/coreprotect/command/PurgeCommand.java
@@ -36,14 +36,14 @@ public class PurgeCommand extends Consumer {
 
     protected static void runCommand(final CommandSender player, boolean permission, String[] args) {
         int resultc = args.length;
-        Location location = CommandHandler.parseLocation(player, args);
-        final Integer[] argRadius = CommandHandler.parseRadius(args, player, location);
-        final List<Integer> argAction = CommandHandler.parseAction(args);
-        final List<Object> argBlocks = CommandHandler.parseRestricted(player, args, argAction);
-        final Map<Object, Boolean> argExclude = CommandHandler.parseExcluded(player, args, argAction);
-        final List<String> argExcludeUsers = CommandHandler.parseExcludedUsers(player, args);
-        final long[] argTime = CommandHandler.parseTime(args);
-        final int argWid = CommandHandler.parseWorld(args, false, false);
+        Location location = CommandParser.parseLocation(player, args);
+        final Integer[] argRadius = CommandParser.parseRadius(args, player, location);
+        final List<Integer> argAction = CommandParser.parseAction(args);
+        final List<Object> argBlocks = CommandParser.parseRestricted(player, args, argAction);
+        final Map<Object, Boolean> argExclude = CommandParser.parseExcluded(player, args, argAction);
+        final List<String> argExcludeUsers = CommandParser.parseExcludedUsers(player, args);
+        final long[] argTime = CommandParser.parseTime(args);
+        final int argWid = CommandParser.parseWorld(args, false, false);
         final List<Integer> supportedActions = Arrays.asList();
         long startTime = argTime[1] > 0 ? argTime[0] : 0;
         long endTime = argTime[1] > 0 ? argTime[1] : argTime[0];
@@ -77,7 +77,7 @@ public class PurgeCommand extends Consumer {
             return;
         }
         if (argWid == -1) {
-            String worldName = CommandHandler.parseWorldName(args, false);
+            String worldName = CommandParser.parseWorldName(args, false);
             Chat.sendMessage(player, new ChatMessage(Phrase.build(Phrase.WORLD_NOT_FOUND, worldName)).build());
             return;
         }
@@ -202,7 +202,7 @@ public class PurgeCommand extends Consumer {
                     }
 
                     if (argWid > 0) {
-                        String worldName = CommandHandler.parseWorldName(args, false);
+                        String worldName = CommandParser.parseWorldName(args, false);
                         Chat.sendGlobalMessage(player, Phrase.build(Phrase.PURGE_STARTED, worldName));
                     }
                     else {
